fix mouse movements

+++ arduino/mouse_utils.h
@@ -11,7 +11,7 @@ class MouseUtils {
 
     void setup();
 
-    void move(int8_t InX, int8_t InY);
+    void move(int InX, int InY);
     void move2(int8_t InX, int8_t InY);
 
     void mouse_click(String event);
